Update test to handle the fact that the server now runs
As the server now runs rather than starts the mvn process does not exit
so need to do it on a separate thread to allow it to do it's thing.
Also, to make the test dependencies cleaner have switched to using Maven
embedded so that this will run on a system without Maven already
installed which also made this platform independent and removed the need
for a Maven check at the start.

In order to check that the build had completed OK and the server was
runner we needed more log checking so wrote some Hamcrest matchers to
keep the tests clean.