Update to latest fixpack and simplify swagger
Update to generate projects for the most recent Liberty release which is 17.0.0.1. Also update the swagger so it doesn't configure a default user account with a default password. Access to the swagger UI now requires the unauthenticated swagger UI. This also allows the use of ldap with the swagger type, previously you'd have to stop using the swagger dependency. Overall this makes things more useful and more secure.